Multipath commands with error messages in stdout fail to parse
This change fixes an issue in find_multipath_device() where the command
output of ‘multipath -l <device>’ would sometimes fail to be parsed if
there were error messages in the stdout string in addition to the
expected output. We will now strip out the error messages before we
attempt to parse the lines.

Change-Id: I3d48debc7d64f6a891ac37d4c2c5ebddfb5b00b0
Closes-Bug: #1380742